+++ tests/testthat/test-brms.R
@@ -900,3 +900,22 @@ test_that("get_modelmatrix", {
     )
   )
 })
+
+# get variance
+test_that("get_variance works", {
+  mdl <- insight::download_model("brms_mixed_9")
+  out <- get_variance(mdl)
+  expect_equal(
+    out,
+    list(
+      var.fixed = 4.91103174480995,
+      var.random = 22.4069708072874,
+      var.residual = 11.3506326649,
+      var.distribution = 11.3506326649,
+      var.dispersion = 0,
+      var.intercept = c(cyl = 22.4069708072874)
+    ),
+    tolerance = 1e-3,
+    ignore_attr = TRUE
+  )
+})