+++ NEWS.md
@@ -24,7 +24,7 @@
 * Fixed issue in `ellipses_info()` when this function was called from `do.call()`.
 
 * Fixed issue with formatting unicode-symbols, where a wrong unicode-character
-  was used for "Omega". Furthermore, Omega2 and Eta2 were correctly converted.
+  was used for "Omega". Furthermore, Omega2 and Eta2 are now correctly converted.
 
 # insight 0.20.2